[py_connector] fix the source-checkout run path and residual split references
Four fixes surfaced by running the e2e harness against the local tree:

- _version_info is stamped into the wheel at build time; a source
  checkout has none, so the connector falls back to dev markers instead
  of failing to import (running from a checkout is a supported mode --
  it is how the e2e harness exercises the code under review).
- data_transfer still passed the deleted kv_stride field to the gather/
  scatter kernels (AttributeError at save time); both call sites drop it.
- the e2e harness resolved worker-role attributes (_data_transfer,
  _tp_rank, ...) on the shell, where the role split moved them to
  ConnectorWorker; properties forward them now. _capture_pending_loaded
  also referenced its caller's meta variable instead of its parameter.
- the harness passes the repository root ahead of any installed wheel
  in the vLLM process's PYTHONPATH, so connector modules resolve to the
  checkout under test; run_e2e moves to servers.py (it orchestrates
  ScenarioEnv) and servers.py imports the utils it consumes.

test_basic passes bit-exact (36/36 blocks) against Qwen2.5-7B-Instruct
on vLLM 0.26.0 with the real manager binary.
