+++ integration_test/vllm_e2e/e2e_lib.py
@@ -11,8 +11,7 @@ from lib_utils import (  # noqa: F401
     block_token_hash, full_block_hashes, wait_for_prefix_cached,
     send_completions, count_captures, wait_for_captures, compare_captures,
     assert_report_ok, make_base_prompts, shared_token_prefix_len,
-    run_e2e,
 )
 from servers import (  # noqa: F401
-    ManagerProcess, VllmServer, ScenarioEnv,
+    ManagerProcess, VllmServer, ScenarioEnv, run_e2e,
 )
