+++ integration_test/vllm_e2e/lib_utils.py
@@ -420,99 +420,3 @@ def shared_token_prefix_len(a: list[int], b: list[int]) -> int:
             break
         n += 1
     return n
-
-
-def run_e2e(scenario: str, tp_size: int, num_prompts: int,
-            preferred_block_size: int, connector_name: str = "VerifyingConnector",
-            expect_verification_failure: bool = False):
-    """Run one full save-then-load verification scenario.
-
-    Full-attention models: prefix caching off, one server across both phases.
-    Hybrid models: prefix caching on (align mode -> per-group block tables), the
-    vLLM server is restarted between phases so phase 2 loads from KVCM instead of
-    hitting the local prefix cache.
-
-    connector_name selects the connector class inside test_connector.py; the
-    mutation meta-test passes "MutatedConnector" and sets
-    expect_verification_failure=True to prove the harness detects an injected
-    off-by-one in the token translation.
-    """
-    import torch  # noqa: F401  (ensure torch importable early for clear errors)
-
-    env = ScenarioEnv(scenario, tp_size=tp_size,
-                      preferred_block_size=preferred_block_size,
-                      connector_name=connector_name)
-    hybrid = env.hybrid
-    capture_dir = env.capture_dir
-    logger.info("scenario=%s model=%s hybrid=%s tp=%d prompts=%d preferred_bs=%d",
-                scenario, MODEL_PATH, hybrid, tp_size, num_prompts,
-                env.preferred_block_size)
-
-    try:
-        env.start_manager()
-        vllm = env.start_vllm(log_suffix="" if not hybrid else "_p1")
-
-        base_prompts = make_base_prompts(num_prompts, hybrid)
-        suffixes = [f" Now answer question {i}: what is 2+2?" for i in range(num_prompts)]
-        phase2_prompts = [p + s for p, s in zip(base_prompts, suffixes)]
-
-        # Compute per-prompt expected block counts from the actual tokenization
-        # so a silently dropped prompt (or block) fails the run.
-        mbs = env.manager_block_size()
-        base_tokens = [tokenize(vllm.base_url(), p) for p in base_prompts]
-        phase2_tokens = [tokenize(vllm.base_url(), p) for p in phase2_prompts]
-        expected_save_blocks = [len(t) // mbs for t in base_tokens]
-        # Loads cover the shared token prefix (the base/suffix boundary token may
-        # re-merge under tokenization, shortening the shared prefix by one).
-        expected_load_blocks = [
-            min(shared_token_prefix_len(b, p2) // mbs, s)
-            for b, p2, s in zip(base_tokens, phase2_tokens, expected_save_blocks)
-        ]
-        assert all(n >= 1 for n in expected_load_blocks), (
-            f"prompts too short to span a manager block (mbs={mbs}): "
-            f"{expected_load_blocks}")
-        logger.info("mbs=%d expected save blocks=%s load blocks=%s",
-                    mbs, expected_save_blocks, expected_load_blocks)
-
-        # ---- Phase 1: fresh prefill -> connector saves -> reference capture.
-        logger.info("phase 1: sending %d fresh prompts", num_prompts)
-        send_completions(vllm.base_url(), base_prompts)
-        wait_for_captures(capture_dir, "ref",
-                          expected=tp_size * sum(expected_save_blocks), timeout=180)
-
-        # The save is committed to the manager asynchronously after the ref
-        # capture (which fires when the save is submitted). Wait until the manager
-        # actually has the prefix, otherwise phase 2 would find no match.
-        for toks, blocks in zip(base_tokens, expected_save_blocks):
-            if not wait_for_prefix_cached(env.manager.manager_uri(), env.instance_id,
-                                          toks, min_blocks=blocks):
-                raise AssertionError("save was not committed to the manager in time")
-
-        # Hybrid models keep prefix caching on, which also populates the local
-        # prefix cache; restart vLLM so phase 2 loads from KVCM, not locally.
-        if hybrid:
-            logger.info("restarting vLLM before phase 2 (clear local prefix cache)")
-            vllm = env.restart_vllm(log_suffix="_p2")
-
-        # ---- Phase 2: same prefix + suffix -> connector loads -> loaded capture.
-        logger.info("phase 2: sending %d prefix+suffix prompts", num_prompts)
-        send_completions(vllm.base_url(), phase2_prompts)
-        wait_for_captures(capture_dir, "loaded",
-                          expected=tp_size * sum(expected_load_blocks), timeout=180)
-
-        report = compare_captures(capture_dir, tp_size)
-        min_matched = tp_size * sum(expected_load_blocks)
-        if expect_verification_failure:
-            try:
-                assert_report_ok(report, min_matched=min_matched)
-            except AssertionError as e:
-                logger.info("verification failed as expected: %s", e)
-                return
-            raise AssertionError(
-                "mutated connector passed KV verification; the harness is blind")
-        assert_report_ok(report, min_matched=min_matched)
-        logger.info("scenario %s PASSED: %s", scenario, json.dumps(
-            {k: v for k, v in report.items() if k not in ("failures", "matched_keys")},
-            default=str))
-    finally:
-        env.stop()
